requirements.txt
torch
torchvision
numpy
opencv-python
pillow
efficientnet_pytorch
resnest
progress
scipy
pandas
ttach